requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Werkzeug==3.1.3
python-dotenv==1.1.1
psycopg2-binary==2.9.10
cloudinary==1.44.1
requests==2.31.0
gunicorn==21.2.0
pytz==2025.2
PyPDF2==3.0.1
Flask-Login==0.6.3
Flask-Compress==1.18
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
supabase==2.22.2
gotrue==2.12.4